+++ radiospectra/exceptions.py
@@ -1,0 +1,9 @@
+__all__ = ["NoSpectrogramInFileError", "SpectraMetaValidationError"]
+
+
+class NoSpectrogramInFileError(Exception):
+    pass
+
+
+class SpectraMetaValidationError(AttributeError):
+    pass