+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.5.0 (2022-10-15)
+==================
+
+Breaking Changes
+----------------
+
+- The old ``Spectrogram`` class has been removed
+- The new ``Spectrogram2`` class has been renamed to ``Spectrogram``.
+
 0.4.0 (2022-05-24)
 ==================
 
@@ -10,11 +19,10 @@ Breaking Changes
 Features
 --------
 
-- Add a new spectrogram class `radiospectra.spectrogram2.spectrogram.BaseSpectrogram` and factory `radiospectra.spectrogram2.spectrogram.SpectrogramFactory` with sources for `~radiospectra.spectrogram2.sources.SWAVESSpectrogram`, `~radiospectra.spectrogram2.sources.RFSSpectrogram`, `~radiospectra.spectrogram2.sources.CALISTOSpectrogram`, `~radiospectra.spectrogram2.sources.EOVSASpectrogram` and `~radiospectra.spectrogram2.sources.RSTNSpectrogram`. (`#44 <https://github.com/sunpy/radiospectra/pull/44>`__)
+- Add a new spectrogram class `radiospectra.spectrogram.spectrogram.BaseSpectrogram` and factory `radiospectra.spectrogram.spectrogram.SpectrogramFactory` with sources for `~radiospectra.spectrogram.sources.SWAVESSpectrogram`, `~radiospectra.spectrogram.sources.RFSSpectrogram`, `~radiospectra.spectrogram.sources.CALISTOSpectrogram`, `~radiospectra.spectrogram.sources.EOVSASpectrogram` and `~radiospectra.spectrogram.sources.RSTNSpectrogram`. (`#44 <https://github.com/sunpy/radiospectra/pull/44>`__)
 - Add `sunpy.net.Fido` clients for `~radiospectra.net.sources.callisto.CALLISTOClient`, `~radiospectra.net.sources.eovsa.EOVSAClient` and `~radiospectra.net.sources.rstn.RSTNClient`. (`#44 <https://github.com/sunpy/radiospectra/pull/44>`__)
-- Improve `~radiospectra.spectrogram2.spectrogram.SpectrogramFactory` input handling more inputs formats data header pairs, files, urls. (`#54 <https://github.com/sunpy/radiospectra/pull/54>`__)
-- Add `sunpy.net.Fido` client `~radiospectra.net.sources.wind.Waves` and spectrogram class `~radiospectra.spectrogram2.sources.WAVESSpectrogram` for WIND/WAVES. (`#54 <https://github.com/sunpy/radiospectra/pull/54>`__)
-
+- Improve `~radiospectra.spectrogram.spectrogram.SpectrogramFactory` input handling more inputs formats data header pairs, files, urls. (`#54 <https://github.com/sunpy/radiospectra/pull/54>`__)
+- Add `sunpy.net.Fido` client `~radiospectra.net.sources.wind.Waves` and spectrogram class `~radiospectra.spectrogram.sources.WAVESSpectrogram` for WIND/WAVES. (`#54 <https://github.com/sunpy/radiospectra/pull/54>`__)
 
 0.3.0 (2021-04-01)
 ==================